add inline eslint rules to work with standard
This is a very temporary solution. I would like to apply small amounts
of changes in my commits and try to keep the original source as close to
the master repo as possible, so future merges will be easy.

I use [standard](https://github.com/feross/standard).
It has strict eslint rules, but the current syntax style doesn't conform
to them. Instead of conforming, I decided to except some, that are
applied by standard.